[level 3] Title: 최고 속도, Time: 578.32 ms, Memory: 191 MB -BaekjoonHub

+++ 프로그래머스/3/468376. 최고 속도/최고 속도.java
@@ -150,7 +150,7 @@ class Solution {
     }
     
     long genKey(int x, int y) {
-        return (long)x * 1_000_000_000L + y;
+        return x * 1_000_000_000L + y;
     }
     
     boolean onRoad(Point point, int[] road) {
